printer_emulator OK
everything works
